// apps/client/src/components/dialogs/dialogs.ts
export enum Dialog {
  CONFIRM_ACTION = 'CONFIRM_ACTION',
  CREATE_CHANNEL = 'CREATE_CHANNEL',
  TEXT_INPUT = 'TEXT_INPUT',
  SERVER_PASSWORD = 'SERVER_PASSWORD',
  SOUNDS = 'SOUNDS',
  ASSIGN_ROLE = 'ASSIGN_ROLE',
  CREATE_INVITE = 'CREATE_INVITE',
  CREATE_CATEGORY = 'CREATE_CATEGORY',
  PLUGIN_LOGS = 'PLUGIN_LOGS',
  PLUGIN_COMMANDS = 'PLUGIN_COMMANDS',
  PLUGIN_SETTINGS = 'PLUGIN_SETTINGS',
  PLUGIN_INSTALL_CONFIRM = 'PLUGIN_INSTALL_CONFIRM',
  DELETE_USER = 'DELETE_USER',
  SEARCH = 'SEARCH',
  WELCOME_PROFILE_SETUP = 'WELCOME_PROFILE_SETUP',
  VOICE_DEVICE_CHECK = 'VOICE_DEVICE_CHECK'
}